brats_challenge/implementation/requirements.txt
torch>=1.6.0.dev20200518+cu92
SimpleITK>=1.2.4
numpy>=1.18.4
scikit-learn>=0.23.1
pandas>=1.0.3
PyYAML>=5.3.1
matplotlib>=3.2.1
scipy>=1.4.1
git+https://github.com/lessw2020/Ranger-Deep-Learning-Optimizer.git
typer
